Version increase since this breaks to some extent the API compatibility for UiElementPropertiesI;
IMPROVEMENT added properties to determine 'selectable' targets and to determine if there are clickable descendants;
This is necessary for apps which do not report elements as 'clickable' but rather as selectable.
The descendants property can be used to reduce the search space if the app reports each parent layer as selectable as well.